SimplexEvaluator
In the evaluation of the triangualtion we do not create a Simplex object
for each simplex, but have a SimplexEvaluator class which is created once
per thread.
This class is splitted from Simplex, which now only has the basic
functionality to compute SupportHyperplanes and volume.

There is still a big optimization potential but it has already a great effect.