Store pyramids of recrusive processing in level 0
All pyramids that are created in the recursive processing of the top cone are stored directly on level 0 (and not on their natural level). There is a check in the program whether this has worked out.

At present the change would be irrelevant, but if we should switch to a pyramid evaluation in which we control the next level, then the change is useful and cannot be forgotten anymore.