Simplification for reduction of candidates
The norm is stored at the end of the vector, so simple push_back and
pop_back operations can be used. Also we can use splice to transfer
candidates more often.
The sorting is done via compare_last, the elements are sorted
considering the last component first. candidates with the same norm
are still sorted totally since it has an positive effect for the
reduction in examples like big and huge.
In the global reduction in the "reducer" list only the scalar products
are stored.