pyramid recursion
1) the notion of level has been adapted to the course of the computation. All pyramids created in the recursive treatment of the top cone are of level 0. If we want we can re-introduce something like natural_level or generation. (Also see 3)

2) The support hyperplanes of pyramids are not fully computed if they are not given back to the mother pyramid (controlled by do_all_hyperplanes). The effect is hardly measurable, but I have inserted it in order to suppress computations that can be suppressed cheaply.

3) If one does not convert do_partial_triangulation into do_triangulation at a certain level, then it runs down to single simplices. This is of course correct, but in general it is suboptimal. The setting at present is as follows:
i: pyramids created in process_pyramid (the recursive treatment of the top cone), are not converted.
ii: pyramids created in evaluate_stored_pyramids get converted if they are of level >= 3. (Level starts at 0!)

Rule i can only be changed in a reasonable way only if we re-introduce the natural level. Rule ii can be changed---all values are legal. The present setting is not so good for A553 (on that level only simplices are left with -N), but good for lo6.