find_new_facets revised
1) The vector "registers" for P,PS,N,NS,Z,ZS eliminated. Instead of the initial lists we make deques --- made the change easier than working with lists throughout.
At leasts this saves one data transformation and may make the memory management more flexible.
The subdivision into the 6 subclasses is not yet parallelized, though it looks as if Christof tried this once.

2) Parallelization of the "multimap".

3) (new) I have eliminated the use of the Boost function count. Usually we only want to know whether something is simplicial, and once we know it is not we can stop counting. This should be realizable especially in examples with many generators.

I hope this version is faster and uses less memory. I have tested it with A553 -s. Gives the correct number of hyperplanes. I have not measured times. For the parallelization in find_new_facets one must keep in mind that it does only help when we are in the top cone. The parallelization over pyramids blocks parallelization in the subordinate functions.