deterministic order in dualisation
In find_new_facets there are double loops of which the outer is parallelized. If they are running in parallel, the inner loops will "fire" new hyperplanes in an unpredictable order into the list Facets. This had two disadvantages:

1) There was no predictable order in the generators when they were produced by dualization. For a while we sorted them lex to make the order independent of the parallelization.

2) New hyperplanes (or generators) that come from the same "old" hyperplane (or generators) in the exterior loop were not kept together ( lex sorting is also bad in this respect).

Now we collect the new objects in separate lists indexed by the exterior loop variable and splice them together at the end.