Introduce default mode
Default mode tries to compute everything but does not give an exception if it
could not succeed.
If something could not be computed which was explicitly asked, we now throw
an exception.
Cone: isComputed(SupportHyperplanes) now is only true if we have the minimal
set of support hyperplanes. In this case the support hyperplanes are given to
the Full_Cone.
Refactoring: ConeProperties object methods to set and check implications