offload: host does not evaluate all lvl 0 pyramids
The host now only evaluates level 0 pyramids until it is interrupted for
the first time. This gives a bigger buffer for offloading. If the level 0
evaluation bound was not reached in the beginning, the host evaluates as
much as it can, to allow the final evaluation to be done on the host.