Add version number
It is configured in the Normaliz CMakeList.txt. CMake will then create a
libnormaliz/version.h which has several defines as well as a getVersion()
function which returns an unsigned int.

This version is also used in the printHeader() of Normaliz. (This part uses
changes submitted by Jerome Benoit <calculus@rezozer.net>.)