ConeProperties goals and options
New methods for ConeProperties to return the subset of activated goals or
options.
The combination DualMode and DefaultMode now means to compute ExtremeRays
and the HilbertBasis (and not try to compute the HilbertSeries, ...).
Normaliz will activate DefaultMode when no goals where specified, esp.
when only the option -d is given.