Provide our own dynamic_bitset implementation
This provides only a subset of boost::dynamic_bitset, based on
those features Normaliz actually uses. This has the advantage
that the resulting code is much shorter and simpler. Also, unused
code often has bugs, while code that was never written is bug free.

Among the features that are left out are:
- custom allocators -- always use the default allocator
- custom block/limb types -- we hardcode unsigned long long
- initialization of bitsets with non-zero data in constructors, resize()
- iterators
- shift operators
- appending, pop_back, push_back
- conversion from/to strings
- is_proper_subset_of
- istream operator>>

Of course any of these could be added in the future, should the need arise