Treat ClassGroup as a complex structure, not a vector
While technically, ClassGroup is implemented as std::vector<Integer>, it is
not a "true vector". Thus getVectorConeProperty() does not support it.

This change ensures that output_type(ClassGroup) returns OutputType::Complex
instead of OutputType::Vector, so that code using the generic property access
APIs does not end up trying to call getVectorConeProperty(ClassGroup).