Let libtool find indirect dependencies
when building -all-static, libtool pulls in the static versions of
libraries that describe them in their .la files. Unfortunately, libtools
is not so widely used, so to use that feature we have to create fake
libtool files.

However, this does not work: the macOS linker cannot link static
archives that define the same object twice. E-ANTIC needs to redefine a
symbol that is broken in some versions of ANTIC. On Linux this is not an
issue since we are hiding that symbol with a version map. Or we can use
--allow-multiple-definition. On macOS there is no such option (anymore)
so we can only work around this with dynamic linking.